requirements.txt
# Core app
streamlit
pipecat
requests
python-dotenv

# Audio recording and playback
sounddevice
simpleaudio
soundfile
scipy
numpy

# Voice Activity Detection
webrtcvad

# Deepgram STT
deepgram-sdk==4.6.0

# Kokoro TTS (and dependencies)
kokoro
transformers
torch
torchaudio
huggingface-hub

